Corrected Microsoft Edge detection
The framerate setting was causing some cameras to not return anything in Edge. The testbed config wasn't intending to send width/height/frameRate constraints to edge, but was because window.adapter doesn't exist.
This may be a less accurate form of detection than whatever the adapter is, but it does at least exist to detect it.