Adding browser check for media stream swap.
Older mobile browsers (particularly Chrome <=83 on Android) require the stop of tracks prior to subsequent requests to `getUserMedia`.
As such, we need to check the browser and make that call on the current existing `MediaStream` going out.
This has somewhat of an undesirable affect in that the broadcast will be empty until `getUserMedia` is resolved. As such, we do the browser check and let more modern browsers handle the swap upon `getUserMedia`.

Fixes RPRO-7627